Add files via upload
This repository contains a simple example of a classification model using TensorFlow to classify Iris flowers. The Iris dataset is a well-known dataset in machine learning, and in this project, we build a neural network model to predict the species of Iris flowers based on their sepal length, sepal width, petal length, and petal width.